fix dark mode, see notes...
This adds a quick executable shim around the darkmode functions,
because neovim was refusing to allow access to parent directories
to source the file with the functions, which is a good thing.

The shim is in the path, so neovim can see and run it no problem.

Some cleanup is needed after this, for the future.